analysis_service/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
openai>=1.99.0
python-multipart>=0.0.20
pandas>=2.3.0
openpyxl>=3.1.0
pypandoc>=1.15
PyYAML>=6.0.2
